feat(github-runner): distribute prepared sandbox images
