+++ src/github-runner/coordinator/src/lib.rs
@@ -1,0 +1,5 @@
+mod coordinator;
+mod github;
+pub mod provider;
+
+pub use coordinator::{CoordinatorArguments, run};
